requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.3
numpy==1.25.1
openpyxl==3.1.2
outcome==1.2.0
pandas==2.0.3
Pillow==10.1.0
psutil==5.9.5
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
wsproto==1.2.0
